Move purge-vault.component.ts to Vault code ownership
