+++ apps/web/src/app/settings/account.component.ts
@@ -4,9 +4,9 @@ import { ModalService } from "@bitwarden/angular/services/modal.service";
 import { UserVerificationService } from "@bitwarden/common/auth/abstractions/user-verification/user-verification.service.abstraction";
 
 import { DeauthorizeSessionsComponent } from "../auth/settings/deauthorize-sessions.component";
+import { PurgeVaultComponent } from "../vault/settings/purge-vault.component";
 
 import { DeleteAccountComponent } from "./delete-account.component";
-import { PurgeVaultComponent } from "./purge-vault.component";
 
 @Component({
   selector: "app-account",
