+++ apps/web/src/app/vault/settings/purge-vault.component.ts
